# docs/swagger.yaml
definitions:
...
  payment.CreateOrderRequest:
    properties:
      amount:
        type: number
      merchant_order_no:
        maxLength: 64
        minLength: 1
        type: string
      notify_url:
        maxLength: 100
        type: string
      order_name:
        maxLength: 64
        type: string
      payment_type:
        type: string
      remark:
        maxLength: 100
        type: string
      return_url:
        maxLength: 100
        type: string
    required:
    - amount
    - order_name
    type: object
  payment.MerchantDistributeRequest:
    properties:
      amount:
        type: number
      out_trade_no:
        maxLength: 64
        minLength: 1
        type: string
      remark:
        maxLength: 100
        type: string
      user_id:
        type: integer
      username:
        type: string
    required:
    - amount
    - user_id
    - username
    type: object
  payment.PayOrderRequest:
    properties:
      order_no:
        type: string
      pay_key:
        maxLength: 6
        type: string
    required:
    - order_no
    - pay_key
    type: object
  payment.QueryMerchantOrderResponse:
    properties:
      addtime:
        example: "2023-12-08 12:00:00"
        type: string
      code:
        example: 1
        type: integer
      endtime:
        example: "2023-12-08 12:05:00"
        type: string
      money:
        example: "10.00"
        type: string
      msg:
        example: 查询订单号成功！
        type: string
      name:
        example: 商品名称
        type: string
      out_trade_no:
        example: M202312080001
        type: string
      pid:
        example: "1001"
        type: string
      status:
        example: 1
        type: integer
      trade_no:
        example: "123456"
        type: string
      type:
        example: epay
        type: string
    type: object
  payment.RefundMerchantOrderResponse:
    properties:
      code:
        example: 1
        type: integer
      msg:
        example: 退款成功
        type: string
    type: object
  payment.RefundOrderRequest:
    properties:
      key:
        type: string
      money:
        type: number
      out_trade_no:
        type: string
      pid:
        type: string
      trade_no:
        type: integer
    required:
    - key
    - money
    - pid
    - trade_no
    type: object
  payment.TransferRequest:
    properties:
      amount:
        type: number
      pay_key:
        maxLength: 6
        type: string
      recipient_id:
        example: "0"
        type: string
      recipient_username:
        type: string
      remark:
        maxLength: 100
        type: string
    required:
    - amount
    - pay_key
    - recipient_id
    - recipient_username
    type: object
...
  util.ResponseAny:
    properties:
      data: {}
      error_msg:
        example: ""
        type: string
    type: object
info:
  contact: {}
  title: LINUX DO Credit
  version: 1.0.0
paths:
  /api.php:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: act
        type: string
      - in: query
        name: key
        required: true
        type: string
      - in: query
        maxLength: 64
        minLength: 1
        name: out_trade_no
        required: true
        type: string
      - in: query
        name: pid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payment.QueryMerchantOrderResponse'
      tags:
      - payment
    post:
      consumes:
      - application/json
      parameters:
      - description: 退款请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/payment.RefundOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/payment.RefundMerchantOrderResponse'
      tags:
      - payment
...
  /api/v1/merchant/payment:
    post:
      consumes:
      - application/json
      parameters:
      - description: 支付订单请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/payment.PayOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      tags:
      - payment
...
  /api/v1/merchant/payment/order:
    get:
      consumes:
      - application/json
      parameters:
      - description: 订单号
        in: query
        name: order_no
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      tags:
      - payment
...
  /api/v1/payment/transfer:
    post:
      consumes:
      - application/json
      parameters:
      - description: 转账请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/payment.TransferRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      tags:
      - payment
...
  /pay/distribute:
    post:
      consumes:
      - application/json
      parameters:
      - description: Basic Auth (base64(client_id:client_secret))
        in: header
        name: Authorization
        required: true
        type: string
      - description: 分发请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/payment.MerchantDistributeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      tags:
      - payment
  /pay/submit.php:
    get:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/payment.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      tags:
      - payment
    post:
      consumes:
      - application/x-www-form-urlencoded
      parameters:
      - description: request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/payment.CreateOrderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.ResponseAny'
      tags:
      - payment
swagger: "2.0"